moved scroll code to settingsOverlay
