+++ src/v2/views/SettingsOverlay/index.tsx
@@ -12,6 +12,7 @@ import { shell, ipcRenderer } from "electron";
 import { app } from "@electron/remote";
 import { preloadService } from "src/v2/machines/preloadMachine";
 
+import * as ScrollArea from "@radix-ui/react-scroll-area";
 import H1 from "src/v2/components/ui/H1";
 import TextField from "src/v2/components/ui/TextField";
 import { Select, SelectOption } from "src/v2/components/ui/Select";
@@ -24,6 +25,46 @@ import AdvancedAction from "./AdvancedAction";
 import DeleteIcon from "@material-ui/icons/Delete";
 import { OverlayProps } from "src/v2/utils/types";
 import toast from "react-hot-toast";
+
+const SCROLLBAR_SIZE = 10;
+
+const ScrollAreaRoot = styled(ScrollArea.Root, {});
+
+const ScrollAreaViewport = styled(ScrollArea.Viewport, {
+  width: "100%",
+  height: "70vh",
+});
+
+const ScrollAreaScrollbar = styled(ScrollArea.Scrollbar, {
+  "&": {
+    display: "flex",
+    userSelect: "none",
+    touchAction: "none",
+    borderRadius: SCROLLBAR_SIZE,
+    margin: `-24px`,
+    transition: "background 160ms ease-out",
+  },
+  "&:hover": { backgroundColor: "$gray80" },
+  '&[data-orientation="vertical"]': { width: SCROLLBAR_SIZE },
+});
+
+const ScrollAreaThumb = styled(ScrollArea.Thumb, {
+  flex: 1,
+  background: "$primary",
+  borderRadius: SCROLLBAR_SIZE,
+  position: "relative",
+  "&::before": {
+    content: '""',
+    position: "absolute",
+    top: "50%",
+    left: "50%",
+    transform: "translate(-50%, -50%)",
+    width: "100%",
+    height: "100%",
+    minWidth: 44,
+    minHeight: 44,
+  },
+});
 
 const Form = styled("form", {
   display: "flex",
@@ -114,112 +155,150 @@ function SettingsOverlay({ onClose, isOpen }: OverlayProps) {
 
   return (
     <OverlayBase isOpen={isOpen} onDismiss={onClose}>
-      <CloseButton onClick={onClose} />
-      <H1 css={{ marginTop: 0 }}>
-        <T _str="Settings" _tags={transifexTags} />
-      </H1>
-      <p>
-        <T _str="" _tags={transifexTags} />
-      </p>
-      <Form onSubmit={handleSubmit(onSubmit, onError)} id="form">
-        <FormSection>
-          <GroupTitle>
-            <T _str="Language" _tags={transifexTags} />
-          </GroupTitle>
-          <span>
-            <T _str="Select the language you want." _tags={transifexTags} />
-          </span>
-          <Controller
-            name="Locale"
-            control={control}
-            render={({ field }) => (
-              <Select {...field}>
-                {languages.map(({ code, localized_name }) => (
-                  <SelectOption key={code} value={code}>
-                    {localized_name}
-                  </SelectOption>
-                ))}
-              </Select>
-            )}
-          />
-          <GroupTitle>
-            <T _str="Connection" _tags={transifexTags} />
-          </GroupTitle>
-          <Checkbox {...register("UseRemoteHeadless")}>
-            <T _str="Use Remote Headless" _tags={transifexTags} />
-          </Checkbox>
-          <GroupTitle>
-            <T _str="Advanced" _tags={transifexTags} />
-          </GroupTitle>
-          {!useRemoteHeadless && (
-            <Checkbox {...register("NoMiner")}>
-              <T _str="Disable Mining" _tags={transifexTags} />
-            </Checkbox>
-          )}
-          <AdvancedAction
-            icon={<DeleteIcon />}
-            onClick={clearCache}
-            text={t("Clear Cache")}
-          />
-          <AdvancedAction
-            link
-            onClick={handleOpenLogPath}
-            text={t("Open Log Folder")}
-          />
-          <AdvancedAction
-            link
-            onClick={handleOpenKeyStorePath}
-            text={t("Open keystore Folder")}
-          />
-        </FormSection>
-        <FormSection>
-          <GroupTitle>
-            <T _str="Root chain path" _tags={transifexTags} />
-          </GroupTitle>
-          <span>
-            <T
-              _str="Path where game chain data is stored."
-              _tags={transifexTags}
-            />
-          </span>
-          <FolderChooser control={control} name="BlockchainStoreDirParent" />
+      <ScrollAreaRoot type="always">
+        <ScrollAreaViewport>
+          <CloseButton onClick={onClose} />
+          <H1 css={{ marginTop: 0 }}>
+            <T _str="Settings" _tags={transifexTags} />
+          </H1>
+          <p>
+            <T _str="" _tags={transifexTags} />
+          </p>
+          <Form onSubmit={handleSubmit(onSubmit, onError)} id="form">
+            <FormSection>
+              <GroupTitle>
+                <T _str="Language" _tags={transifexTags} />
+              </GroupTitle>
+              <span>
+                <T _str="Select the language you want." _tags={transifexTags} />
+              </span>
+              <Controller
+                name="Locale"
+                control={control}
+                render={({ field }) => (
+                  <Select {...field}>
+                    {languages.map(({ code, localized_name }) => (
+                      <SelectOption key={code} value={code}>
+                        {localized_name}
+                      </SelectOption>
+                    ))}
+                  </Select>
+                )}
+              />
+              <GroupTitle>
+                <T _str="Connection" _tags={transifexTags} />
+              </GroupTitle>
+              <Checkbox {...register("UseRemoteHeadless")}>
+                <T _str="Use Remote Headless" _tags={transifexTags} />
+              </Checkbox>
+              <GroupTitle>
+                <T _str="Advanced" _tags={transifexTags} />
+              </GroupTitle>
+              {!useRemoteHeadless && (
+                <Checkbox {...register("NoMiner")}>
+                  <T _str="Disable Mining" _tags={transifexTags} />
+                </Checkbox>
+              )}
+              <AdvancedAction
+                icon={<DeleteIcon />}
+                onClick={clearCache}
+                text={t("Clear Cache")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenLogPath}
+                text={t("Open Log Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+              <AdvancedAction
+                link
+                onClick={handleOpenKeyStorePath}
+                text={t("Open keystore Folder")}
+              />
+            </FormSection>
+            <FormSection>
+              <GroupTitle>
+                <T _str="Root chain path" _tags={transifexTags} />
+              </GroupTitle>
+              <span>
+                <T
+                  _str="Path where game chain data is stored."
+                  _tags={transifexTags}
+                />
+              </span>
+              <FolderChooser
+                control={control}
+                name="BlockchainStoreDirParent"
+              />
 
-          <GroupTitle>
-            <T _str="Chain store directory name" _tags={transifexTags} />
-          </GroupTitle>
-          <TextField
-            label={t("Chain store directory name")}
-            readOnly
-            {...register("BlockchainStoreDirName")}
-          />
+              <GroupTitle>
+                <T _str="Chain store directory name" _tags={transifexTags} />
+              </GroupTitle>
+              <TextField
+                label={t("Chain store directory name")}
+                readOnly
+                {...register("BlockchainStoreDirName")}
+              />
 
-          <GroupTitle>
-            <T _str="Send Information" _tags={transifexTags} />
-          </GroupTitle>
-          <Checkbox {...register("Mixpanel")}>
-            <T _str="Send anonymous usage information" _tags={transifexTags} />
-          </Checkbox>
-          <Checkbox {...register("Sentry")}>
-            <T _str="Report errors and bugs" _tags={transifexTags} />
-          </Checkbox>
+              <GroupTitle>
+                <T _str="Send Information" _tags={transifexTags} />
+              </GroupTitle>
+              <Checkbox {...register("Mixpanel")}>
+                <T
+                  _str="Send anonymous usage information"
+                  _tags={transifexTags}
+                />
+              </Checkbox>
+              <Checkbox {...register("Sentry")}>
+                <T _str="Report errors and bugs" _tags={transifexTags} />
+              </Checkbox>
 
-          <GroupTitle>
-            <T _str="Debugging" _tags={transifexTags} />
-          </GroupTitle>
-          <Checkbox {...register("PreferLegacyInterface")}>
-            <T _str="Use Legacy Interface" _tags={transifexTags} />
-          </Checkbox>
-        </FormSection>
-      </Form>
-      <Button
-        type="submit"
-        form="form"
-        disabled={!isDirty}
-        centered
-        variant="primary"
-      >
-        <T _str="Save" _tags={transifexTags} />
-      </Button>
+              <GroupTitle>
+                <T _str="Debugging" _tags={transifexTags} />
+              </GroupTitle>
+              <Checkbox {...register("PreferLegacyInterface")}>
+                <T _str="Use Legacy Interface" _tags={transifexTags} />
+              </Checkbox>
+            </FormSection>
+          </Form>
+          <Button
+            type="submit"
+            form="form"
+            disabled={!isDirty}
+            centered
+            variant="primary"
+          >
+            <T _str="Save" _tags={transifexTags} />
+          </Button>
+        </ScrollAreaViewport>
+        <ScrollAreaScrollbar orientation="vertical">
+          <ScrollAreaThumb />
+        </ScrollAreaScrollbar>
+      </ScrollAreaRoot>
     </OverlayBase>
   );
 }
